requirements.txt
python-decouple
beautifulsoup4
selenium
pyodbc